+++ src/Fantomas.Core/CodePrinter.fs
@@ -1591,6 +1591,11 @@ let genExpr (e: Expr) =
         |> genNode node
     | Expr.IndexFromEnd node -> !- "^" +> genExpr node.Expr |> genNode node
     | Expr.Typar node -> genSingleTextNode node
+    | Expr.DotLambda node ->
+        genSingleTextNode node.Underscore
+        +> genSingleTextNode node.Dot
+        +> genExpr node.Expr
+        |> genNode node
 
 let genQuoteExpr (node: ExprQuoteNode) =
     genSingleTextNode node.OpenToken
@@ -2469,6 +2474,7 @@ let genAppWithLambda sep (node: ExprAppWithLambdaNode) =
 
 let sepSpaceBeforeParenInFuncInvocation (functionExpr: Expr) (argExpr: Expr) ctx =
     match functionExpr, argExpr with
+    | Expr.DotLambda _, _ -> ctx
     | Expr.Constant _, _ -> sepSpace ctx
     | ParenExpr _, _ -> sepSpace ctx
     | UppercaseExpr, ParenExpr _ -> onlyIf ctx.Config.SpaceBeforeUppercaseInvocation sepSpace ctx
@@ -3166,6 +3172,11 @@ let genType (t: Type) =
         |> genNode node
     | Type.LongIdentApp node ->
         genType node.AppType +> sepDot +> genIdentListNode node.LongIdent
+        |> genNode node
+    | Type.Intersection node ->
+        col sepSpace node.TypesAndSeparators (function
+            | Choice1Of2 t -> genType t
+            | Choice2Of2 amp -> genSingleTextNode amp)
         |> genNode node
 
 let genSynTupleTypeSegments (path: Choice<Type, SingleTextNode> list) =